move server test and fixed failing 'mvn install'
