+++ server/src/test/java/org/robovm/junit/server/RunnerClass.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright (C) 2014 RoboVM AB
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+package org.robovm.junit.server;
+
+import static org.junit.Assert.*;
+
+import org.junit.Test;
+
+public class RunnerClass {
+
+    @Test
+    public void testSuccessfulTest1() {
+        assertFalse(1 == 2);
+    }
+
+    @Test
+    public void testSuccessfulTest2() {
+        assertNotEquals(0, System.currentTimeMillis());
+    }
+
+    @Test
+    public void testShouldFail() {
+        assertTrue("1 == 2", 1 == 2);
+    }
+}
